fix: import Hero section Framer Motion's components on client-side
