+++ src/components/DataRoom/Hero/styles.module.css
@@ -4,7 +4,7 @@
 }
 
 .stickyContainer {
-  background-color: #000;
+  background-color: var(--mui-palette-background-default);
   display: flex;
   flex-direction: column;
   gap: 32px;
@@ -18,6 +18,19 @@
   max-width: calc(100dvw - 56px);
   height: calc(100dvh - 128px);
   overflow: hidden;
+  margin: auto;
+}
+
+.content {
+  display: flex;
+  flex-direction: column;
+  gap: 32px;
+  margin-bottom: 48px;
+  z-index: 20;
+}
+
+.text {
+  width: 70%;
   margin: auto;
 }
 
@@ -39,26 +52,12 @@
   margin: auto;
 }
 
-.content {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  gap: 32px;
-  margin-bottom: 48px;
-  z-index: 20;
-}
-
 .asciiContainer {
   position: absolute;
   top: 0;
   left: 0;
   width: 100%;
   height: 100%;
-}
-
-.text {
-  width: 70%;
-  margin: auto;
 }
 
 @media (max-width: 900px) {
