+++ src/content/dataroom.json
@@ -8,7 +8,7 @@
   {
     "component": "DataRoom/Hero",
     "title": "Welcome to Safe Data Room",
-    "text": "Explore Mind Bending Sats from Safe Universe"
+    "text": "Explore Mind Bending Stats from Safe Universe"
   },
   {
     "component": "DataRoom/WorldGDP",
